First Milestone
Features an entire playable of Chapter 1 Quiz. Press a word or a clue
to focus. On pc you can type when word is in focus, on mobile you must
press the keyboard button. Once the puzzle has been filled, press
solve. A highscore table will come up.
notes -
Keyboard button will not be visible on pc
Background images are missing
Exiting the level will cause it to crash(because there is no level to
load)